polymarket_bot/requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
